requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amazon-textract-caller==0.2.3
amazon-textract-response-parser==1.0.2
amazon-textract-textractor==1.7.12
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
boto3==1.34.116
botocore==1.34.116
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crashtest==0.4.1
dataclasses-json==0.6.6
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
editdistance==0.8.1
fastapi==0.110.3
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
googleapis-common-protos==1.63.0
groq==0.8.0
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kubernetes==29.0.0
langchain==0.2.1
langchain-cli==0.0.24
langchain-community==0.2.1
langchain-core==0.2.3
langchain-groq==0.1.4
langchain-text-splitters==0.2.0
langserve==0.2.1
langsmith==0.1.67
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
posthog==3.5.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.18.3
Pygments==2.18.0
pypdf==4.2.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
rapidfuzz==3.9.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sse-starlette==1.8.2
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
trove-classifiers==2024.5.22
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
xattr==1.1.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.19.1
